requirements.test.txt
pytest
pytest-cov
pytest-django
pytest-random-order
